Push for Menu, Register and Login.
Login and Register have been changed.
Menu has been added, but it isn't routed anywhere.
The other pages will be in "/cedar 2" for the Front home page and Menu html's.
The other pages need to be brought to the main folder for the Cedar Market webpage and routed through. Other than that, it should be functional if routed correctly.
(I don't know if something is wrong with the Menu page, however I couldn't get it to run nor route through and I don't know if its because of the code or routing problems.)